push version with checks on roles parents and childs

+++ src/main/java/com/example/examplequerydslspringdatajpamaven/service/GeofenceServiceImpl.java
@@ -13,6 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
+import com.example.examplequerydslspringdatajpamaven.entity.CustomDeviceList;
 import com.example.examplequerydslspringdatajpamaven.entity.Driver;
 import com.example.examplequerydslspringdatajpamaven.entity.Geofence;
 import com.example.examplequerydslspringdatajpamaven.entity.User;
@@ -75,8 +76,45 @@ public class GeofenceServiceImpl extends RestServiceController implements Geofen
 					}
 				}
 				if(user.getDelete_date() == null) {
-					geofences = geofenceRepository.getAllGeofences(id,offset,search);
-					Integer size=geofenceRepository.getAllGeofencesSize(id);
+					
+				    userServiceImpl.resetChildernArray();
+				    if(user.getAccountType() == 4) {
+						 Set<User> parentClients = user.getUsersOfUser();
+						 if(parentClients.isEmpty()) {
+							
+							 getObjectResponse = new GetObjectResponse(HttpStatus.NOT_FOUND.value(), "you cannot get geofences of this user",null);
+							 logger.info("************************ getAllUserDevices ENDED ***************************");
+							return  ResponseEntity.status(404).body(getObjectResponse);
+						 }else {
+							 User parentClient = new User() ;
+							 for(User object : parentClients) {
+								 parentClient = object;
+							 }
+							 List<Long>usersIds= new ArrayList<>();
+							 usersIds.add(parentClient.getId());
+							 geofences = geofenceRepository.getAllGeofences(usersIds,offset,search);
+							 Integer size=geofenceRepository.getAllGeofencesSize(usersIds);
+							getObjectResponse= new GetObjectResponse(HttpStatus.OK.value(), "Success",geofences,size);
+							logger.info("************************ getAllUserGeofences ENDED ***************************");
+							return  ResponseEntity.ok().body(getObjectResponse);
+						 }
+					 }
+				    List<User>childernUsers = userServiceImpl.getActiveAndInactiveChildern(id);
+					 List<Long>usersIds= new ArrayList<>();
+					 if(childernUsers.isEmpty()) {
+						 usersIds.add(id);
+					 }
+					 else {
+						 usersIds.add(id);
+						 for(User object : childernUsers) {
+							 usersIds.add(object.getId());
+						 }
+					 }
+
+					
+					
+				    geofences = geofenceRepository.getAllGeofences(usersIds,offset,search);
+					Integer size=geofenceRepository.getAllGeofencesSize(usersIds);
 					getObjectResponse= new GetObjectResponse(HttpStatus.OK.value(), "Success",geofences,size);
 					logger.info("************************ getAllUserGeofences ENDED ***************************");
 					return  ResponseEntity.ok().body(getObjectResponse);
@@ -102,7 +140,7 @@ public class GeofenceServiceImpl extends RestServiceController implements Geofen
 	}
 
 	@Override
-	public ResponseEntity<?> getGeofenceById(String TOKEN,Long geofenceId) {
+	public ResponseEntity<?> getGeofenceById(String TOKEN,Long geofenceId,Long userId) {
 		logger.info("************************ getGeofenceById STARTED ***************************");
 
 		List<Geofence> geofences = new ArrayList<Geofence>();
@@ -115,18 +153,56 @@ public class GeofenceServiceImpl extends RestServiceController implements Geofen
 		{
 			return super.checkActive(TOKEN);
 		}
+		if(userId == 0) {
+       	 getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "loggedUser id is required",null);
+			 return  ResponseEntity.badRequest().body(getObjectResponse);
+       }
+       User loggedUser = userServiceImpl.findById(userId);
+       if(loggedUser == null) {
+       	getObjectResponse= new GetObjectResponse(HttpStatus.NOT_FOUND.value(), "loggedUser is not Found",geofences);
+			return  ResponseEntity.status(404).body(getObjectResponse);
+       }
 		if(geofenceId != 0) {
 			
 			Geofence geofence=geofenceRepository.findOne(geofenceId);
 
 			if(geofence != null) {
 				if(geofence.getDelete_date() == null) {
-					
+					boolean isParent = false;
+					if(loggedUser.getAccountType() == 4) {
+						Set<User> clientParents = loggedUser.getUsersOfUser();
+						if(clientParents.isEmpty()) {
+							getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "you are not allowed to get this geofence",null);
+							 return  ResponseEntity.badRequest().body(getObjectResponse);
+						}else {
+							User parent = null;
+							for(User object : clientParents) {
+								parent = object ;
+							}
+							Set<User>geofneceParents = geofence.getUserGeofence();
+							if(geofneceParents.isEmpty()) {
+								getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "you are not allowed to get this geofnece",null);
+								 return  ResponseEntity.badRequest().body(getObjectResponse);
+							}else {
+								for(User parentObject : geofneceParents) {
+									if(parentObject.getId() == parent.getId()) {
+										isParent = true;
+										break;
+									}
+								}
+							}
+						}
+					}
+					if(!checkIfParent(geofence , loggedUser) && ! isParent) {
+						getObjectResponse = new GetObjectResponse( HttpStatus.BAD_REQUEST.value(), "you are not allowed to get this driver ",null);
+						logger.info("************************ editDevice ENDED ***************************");
+						return ResponseEntity.badRequest().body(getObjectResponse);
+					}
 					geofences.add(geofence);
 					getObjectResponse= new GetObjectResponse(HttpStatus.OK.value(), "Success",geofences);
-					logger.info("************************ getGeofenceById ENDED ***************************");
+					logger.info("************************ getDriverById ENDED ***************************");
 					return  ResponseEntity.ok().body(getObjectResponse);
-
+					
 				}
 				else {
 					getObjectResponse= new GetObjectResponse(HttpStatus.NOT_FOUND.value(), "This Geofence ID is not Found",geofences);
@@ -192,13 +268,45 @@ public class GeofenceServiceImpl extends RestServiceController implements Geofen
 			if(geofence != null) {
 				
 				if(geofence.getDelete_date()==null) {
+					 boolean isParent = false;
+					 if(user.getAccountType() == 4) {
+						 Set<User> parentClients = user.getUsersOfUser();
+						 if(parentClients.isEmpty()) {
+							 getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "you are not allowed to delete this geofnece",geofences);
+							 return  ResponseEntity.badRequest().body(getObjectResponse);
+						 }else {
+							 User parent = null;
+							 for(User object : parentClients) {
+								 parent = object;
+							 }
+							 Set<User>geofneceParent = geofence.getUserGeofence();
+							 if(geofneceParent.isEmpty()) {
+								 getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "you are not allowed to delete this geofnece",geofences);
+								 return  ResponseEntity.badRequest().body(getObjectResponse);
+							 }else {
+								 for(User parentObject : geofneceParent) {
+									 if(parentObject.getId() == parent.getId()) {
+										 isParent = true;
+										 break;
+									 }
+								 }
+							 }
+						 }
+					 }
+					 if(!checkIfParent(geofence , user) && ! isParent) {
+							getObjectResponse = new GetObjectResponse( HttpStatus.BAD_REQUEST.value(), "you are not allowed to delete this geofnece ",geofences);
+							logger.info("************************ deleteGeofence ENDED ***************************");
+							return ResponseEntity.badRequest().body(getObjectResponse);
+						}
+						geofenceRepository.deleteGeofence(geofenceId,currentDate);
+						geofenceRepository.deleteGeofenceId(geofenceId);
+						geofenceRepository.deleteGeofenceDeviceId(geofenceId);
+						getObjectResponse= new GetObjectResponse(HttpStatus.OK.value(), "Deleted Successfully",geofences);
+						logger.info("************************ deleteGeofence ENDED ***************************");
+						return  ResponseEntity.ok().body(getObjectResponse);
 					
-					geofenceRepository.deleteGeofence(geofenceId,currentDate);
-					geofenceRepository.deleteGeofenceId(geofenceId);
-					geofenceRepository.deleteGeofenceDeviceId(geofenceId);
-					getObjectResponse= new GetObjectResponse(HttpStatus.OK.value(), "Deleted Successfully",geofences);
-					logger.info("************************ deleteGeofence ENDED ***************************");
-					return  ResponseEntity.ok().body(getObjectResponse);
+					
+					
 
 				}
 				else {
@@ -371,6 +479,37 @@ public class GeofenceServiceImpl extends RestServiceController implements Geofen
 						 Geofence geofneceCheck = getById(geofence.getId());
 						if(geofneceCheck != null) {
 							if(geofneceCheck.getDelete_date() == null) {
+								boolean isParent = false;
+								
+								if(user.getAccountType() == 4) {
+									Set<User>parentClient = user.getUsersOfUser();
+									if(parentClient.isEmpty()) {
+										 getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "this user is not allowed to edit geofnece",geofences);
+										 return  ResponseEntity.badRequest().body(getObjectResponse);
+									}
+									User parent = null;
+									for(User object : parentClient) {
+										parent = object ;
+									}
+									Set<User>geofenceParent = geofneceCheck.getUserGeofence();
+									if(geofenceParent.isEmpty()) {
+										 getObjectResponse = new GetObjectResponse(HttpStatus.BAD_REQUEST.value(), "this user is not allowed to edit geofnece",geofences);
+										 return  ResponseEntity.badRequest().body(getObjectResponse);
+									}
+									for(User parentObject : geofenceParent) {
+										if(parentObject.getId() == parent.getId()) {
+											isParent = true;
+											break;
+										}
+									}
+								}
+								if(!checkIfParent(geofneceCheck , user) && ! isParent) {
+									getObjectResponse = new GetObjectResponse( HttpStatus.BAD_REQUEST.value(), "you are not allowed to edit this geofence ",null);
+									logger.info("************************ editGeofnece ENDED ***************************");
+									return ResponseEntity.badRequest().body(getObjectResponse);
+								}
+								
+								
 								if(geofence.getName()== null || geofence.getType()== null
 										   || geofence.getArea() == null || geofence.getName()== "" || geofence.getType()== ""
 										   || geofence.getArea() == "") {
@@ -539,5 +678,39 @@ logger.info("************************ getAllUserGeofences STARTED **************
 		}
 	}
 	
+	 public Boolean checkIfParent(Geofence geofnece , User loggedUser) {
+		   Set<User> geofenceParent = geofnece.getUserGeofence();
+		   if(geofenceParent.isEmpty()) {
+			  
+			   return false;
+		   }else {
+			   User parent = null;
+			   for (User object : geofenceParent) {
+				   parent = object;
+			   }
+			   if(parent.getId() == loggedUser.getId()) {
+				   return true;
+			   }
+			   if(parent.getAccountType() == 1) {
+				   if(parent.getId() == loggedUser.getId()) {
+					   return true;
+				   }
+			   }else {
+				   List<User> parents = userServiceImpl.getAllParentsOfuser(parent, parent.getAccountType());
+				   if(parents.isEmpty()) {
+					   
+					   return false;
+				   }else {
+					   for(User object :parents) {
+						   if(object.getId() == loggedUser.getId()) {
+							   return true;
+						   }
+					   }
+				   }
+			   }
+			  
+		   }
+		   return false;
+	   }
 
 }
